fix: main.pyに定義していたschemaを切り出し
